Folding in the agent memory model changes.
