+++ agents_playground/agents/spec/agent_characteristics.py
@@ -2,7 +2,7 @@
 from typing import NamedTuple
 
 from agents_playground.agents.spec.agent_identity_spec import AgentIdentityLike
-from agents_playground.agents.spec.agent_memory_spec import AgentMemoryLike
+from agents_playground.agents.spec.agent_memory_model import AgentMemoryLike
 from agents_playground.agents.spec.agent_movement_attributes import AgentMovementAttributes
 from agents_playground.agents.spec.agent_physicality_spec import AgentPhysicalityLike
 from agents_playground.agents.spec.agent_position_spec import AgentPositionLike
