+++ agents_playground/agents/spec/agent_spec.py
@@ -6,11 +6,11 @@ from __future__ import annotations
 
 from typing import Dict, List, Protocol
 from agents_playground.agents.spec.agent_characteristics import AgentCharacteristics
+from agents_playground.agents.spec.agent_memory_spec import AgentMemoryLike
 
 from agents_playground.agents.spec.agent_system import AgentSystemLike
 from agents_playground.agents.spec.agent_identity_spec import AgentIdentityLike
 from agents_playground.agents.spec.agent_life_cycle_phase import AgentLifeCyclePhase
-from agents_playground.agents.spec.agent_memory_spec import AgentMemoryLike
 from agents_playground.agents.spec.agent_movement_attributes import AgentMovementAttributes
 from agents_playground.agents.spec.agent_physicality_spec import AgentPhysicalityLike
 from agents_playground.agents.spec.agent_position_spec import AgentPositionLike
@@ -25,13 +25,13 @@ from agents_playground.spatial.vector import Vector
 class AgentLike(FrameTick, Protocol):
   """Behaves like an autonomous agent."""
   agent_state: AgentStateLike        # The internal state of the agent.
-  internal_systems: AgentSystemLike      # The subsystems that compose the agent.
+  internal_systems: AgentSystemLike  # The subsystems that compose the agent.
   identity: AgentIdentityLike        # All of the agent's IDs.
   physicality: AgentPhysicalityLike  # The agent's physical attributes.
   position: AgentPositionLike        # All the attributes related to where the agent is.     
   movement: AgentMovementAttributes  # Attributes used for movement.
   style: AgentStyleLike              # Define's the agent's look.
-  memory: AgentMemoryLike            # The memory store for the agent.
+  memory: AgentMemoryLike           # The memory store for the agent.
 
   """
   Thoughts:
